+++ Sources/cmdshelf/Functions.swift
@@ -1,4 +1,5 @@
 import Foundation
+import Poxis
 import Reporter
 
 @discardableResult
@@ -41,4 +42,22 @@ func shellOut(to: String, argument: String? = nil, shouldPrintStdout: Bool = tru
     inPipe.fileHandleForWriting.writeabilityHandler = nil
     #endif
     return process.terminationStatus
+}
+
+/// TODO: improve error handling
+@discardableResult
+func spawnPager(cmdString: String) -> Int32 {
+    errno = 0
+    guard let fpin = poxis_popen(cmdString, "r") else {
+        return 1
+    }
+    guard let fpout = poxis_popen("more", "w") else {
+        return 1
+    }
+    var line: [Int8] = []
+    while fgets(&line, 4096, fpin) != nil {
+        fputs(&line, fpout)// == EOF then error
+    }
+    poxis_pclose(fpout)
+    return poxis_pclose(fpin)
 }